Refactor server UI serving helpers
